@@ -20,10 +20,8 @@
     ],
     "minimum-stability": "stable",
     "require": {
-        "php": ">=5.4.0"
-    },
-    "require-dev": {
-        "phpunit/phpunit": "~4.0"
+        "php": ">=7.2.0",
+        "phpunit/phpunit": "~8.0"
     },
     "autoload": {
         "files": ["src/Inflection.php"]
